+++ src/Csv/CsvGenerator.fs
@@ -39,7 +39,7 @@ module internal CsvTypeBuilder =
     
     let rowType = ProvidedTypeDefinition("Row", Some rowErasedType, HideObjectMethods = true)
 
-    let ctor = ProvidedConstructor([ for field in fields -> field.Param ], InvokeCode = (fun args -> Expr.NewTuple(args) ))
+    let ctor = ProvidedConstructor([ for field in fields -> field.Param ], InvokeCode = fun args -> Expr.NewTuple(args))
     rowType.AddMember ctor
     
     // Each property of the generated row type will simply be a tuple get