+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 * Added constants for more HTTP methods.
 * Added fix for `thead` element without nested `tr` element.
 * Improved global inference in XmlProvider.
+* Write API for CsvProvider.
 
 #### 2.1.1 - December 24 2014
 * Add SkipRows parameter to CsvProvider.